feat(Campground): move ImageGallery to separate component
